refactor: address PR #35 review findings
- Extract WelcomeExperiencePolicy into its own file (Services/WelcomeExperiencePolicy.swift)
- Rename SettingsWindowPresentationStateTests class to match its file name
- Add comment explaining URL Apple Event ordering assumption in handleURLEvent
- Replace XCTUnwrap with XCTSkip in AppSwitcherTests when no regular app is running
